我修改了一下Manifest.xml文件，拉取的时候可以检查一下是否有冲突，在本机测试功能都正常，将推送改成了极光推送

+++ app/src/main/java/com/example/dell/chihuobao/fragment/BarChartFragment.java
@@ -185,7 +185,7 @@ public class BarChartFragment extends BaseFragment implements OnChartGestureList
     private void setData() {
 
         RequestParams params = new RequestParams(BARCHART_URL);
-        params.addQueryStringParameter("shopId",Double.parseDouble(hashMap.get("id").toString())+"");
+        params.addQueryStringParameter("shopId",(int)Double.parseDouble(hashMap.get("id").toString())+"");
         x.http().post(params, new Callback.CommonCallback<String>() {
             @Override
             public void onSuccess(String result) {
